+++ lib/flutter_voximplant.dart
@@ -6,6 +6,7 @@ import 'dart:io';
 import 'dart:async';
 import 'package:flutter/services.dart';
 
+part 'src/log.dart';
 part 'src/voximplant.dart';
 part 'src/client/client.dart';
 part 'src/client/client_config.dart';
@@ -13,4 +14,3 @@ part 'src/client/client_events.dart';
 part 'src/call/call.dart';
 part 'src/call/endpoint.dart';
 part 'src/hardware/audio_device_manager.dart';
-