+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+# Changelog
+
+## 1.1.0
+
+* Add API for CallKit integration on iOS platform
+* Update Android and iOS platform code to use Voximplant Android SDK 2.14.1
+  and Voximplant iOS SDK 2.25.1
+
 ## 1.0.0
+
 * Audio call functionality